requirements.txt
# Offroad Autonomy — Semantic Segmentation
# Install with:  pip install -r requirements.txt

# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Data & augmentation
albumentations>=1.3.1
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Experiment tracking
tensorboard>=2.14.0

# Config
PyYAML>=6.0

# Progress bars & utilities
tqdm>=4.65.0

# Optional — nice to have
matplotlib>=3.7.0      # for quick notebook plots
scikit-learn>=1.3.0    # for class weight computation
